instrumentation/python/traces/sql/app/requirements.txt
flask
opentelemetry-exporter-otlp
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask>=0.8b0
opentelemetry-instrumentation-requests>=0.8b0
opentelemetry-instrumentation-sqlite3>=0.11b0
requests